Покрытие тестами по меню adopt на 100%
