+++ src/test/java/pro/sky/AnimalShelter/handlers/adoptionMenuHandlers/KittyHomeSetupRecommendationHandlerTest.java
@@ -1,4 +1,4 @@
-package pro.sky.AnimalShelter.adoptionMenuHandlers;
+package pro.sky.AnimalShelter.handlers.adoptionMenuHandlers;
 
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.model.Chat;
@@ -20,6 +20,7 @@ import pro.sky.AnimalShelter.utils.CommonUtils;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
+import static pro.sky.AnimalShelter.utils.MessagesBot.HOME_SETUP_RECOMMENDATION_CAT_TEXT;
 import static pro.sky.AnimalShelter.utils.MessagesBot.KITTY_HOME_SETUP_RECOMMENDATION_TEXT;
 
 @ExtendWith(MockitoExtension.class)
@@ -52,6 +53,17 @@ public class KittyHomeSetupRecommendationHandlerTest {
         lenient().when(message.chat()).thenReturn(chat);
         lenient().when(chat.id()).thenReturn(123L);
     }
+    @Test
+    @DisplayName("Проверяет, что при выполнении команды /kitty_home_setup_recommendation, если текущее состояние чата " +
+            "(chatId) равно /ADOPT,  будет отправлено правильное сообщение в чат с указанным chatId")
+    public void testHandleCatStateSendCorrectMessage() {
+        Long chatId = 123L;
+        when(chatStateService.getCurrentStateByChatId(123L)).thenReturn(BotCommand.ADOPT);
+        kittyHomeSetupRecommendationHandler.handle(update);
+        SendMessage message = new SendMessage(chatId, KITTY_HOME_SETUP_RECOMMENDATION_TEXT);
+        telegramBot.execute(message);
+        verify(telegramBot, times(1)).execute(message);
+    }
 
     @Test
     @DisplayName("Проверяет, что при вызове метода handle класса KittyHomeSetupRecommendationHandler " +
@@ -62,7 +74,7 @@ public class KittyHomeSetupRecommendationHandlerTest {
 
         SendMessage message = new SendMessage(chatId, KITTY_HOME_SETUP_RECOMMENDATION_TEXT);
         telegramBot.execute(message);
-        verify(telegramBot, times(1)).execute(any(SendMessage.class));
+        verify(telegramBot, times(1)).execute(message);
     }
 
     @Test
