+++ src/test/java/pro/sky/AnimalShelter/handlers/adoptionMenuHandlers/RefusalReasonHandlerTest.java
@@ -1,4 +1,4 @@
-package pro.sky.AnimalShelter.adoptionMenuHandlers;
+package pro.sky.AnimalShelter.handlers.adoptionMenuHandlers;
 
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.model.Chat;
@@ -13,18 +13,18 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import pro.sky.AnimalShelter.enums.BotCommand;
-import pro.sky.AnimalShelter.handlers.adoptionMenuHandlers.TransportationRecommendationHandler;
+import pro.sky.AnimalShelter.handlers.adoptionMenuHandlers.RefusalReasonHandler;
 import pro.sky.AnimalShelter.service.ChatStateService;
 import pro.sky.AnimalShelter.utils.CommonUtils;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
-import static pro.sky.AnimalShelter.utils.MessagesBot.TRANSPORTATION_RECOMMENDATION_CAT_TEXT;
-import static pro.sky.AnimalShelter.utils.MessagesBot.TRANSPORTATION_RECOMMENDATION_DOG_TEXT;
+import static pro.sky.AnimalShelter.utils.MessagesBot.KITTY_HOME_SETUP_RECOMMENDATION_TEXT;
+import static pro.sky.AnimalShelter.utils.MessagesBot.REFUSAL_REASON_TEXT;
 
 @ExtendWith(MockitoExtension.class)
-public class TransportationRecommendationHandlerTest {
+public class RefusalReasonHandlerTest {
 
     @Mock
     private ChatStateService chatStateService;
@@ -45,7 +45,7 @@ public class TransportationRecommendationHandlerTest {
     private Chat chat;
 
     @InjectMocks
-    private TransportationRecommendationHandler transportationRecommendationHandler;
+    private RefusalReasonHandler refusalReasonHandler;
 
     @BeforeEach
     public void setUp() {
@@ -53,59 +53,58 @@ public class TransportationRecommendationHandlerTest {
         lenient().when(message.chat()).thenReturn(chat);
         lenient().when(chat.id()).thenReturn(123L);
     }
-
     @Test
-    @DisplayName("Проверяет, что при вызове метода handle класса TransportationRecommendationHandler " +
-            "в состоянии \"Рекомендации по транспортировке собаки\" отправляется сообщение с текстом " +
-            "в чат с заданным chatId.")
-    public void testTransportationRecommendationHandlerDog() {
+    @DisplayName("Проверяет, что при выполнении команды /refusal_reason, если текущее состояние чата " +
+            "(chatId) равно /ADOPT,  будет отправлено правильное сообщение в чат с указанным chatId")
+    public void testHandleCatStateSendCorrectMessage() {
         Long chatId = 123L;
-
-        SendMessage message = new SendMessage(chatId, TRANSPORTATION_RECOMMENDATION_DOG_TEXT);
+        when(chatStateService.getCurrentStateByChatId(123L)).thenReturn(BotCommand.ADOPT);
+        refusalReasonHandler.handle(update);
+        SendMessage message = new SendMessage(chatId, REFUSAL_REASON_TEXT);
         telegramBot.execute(message);
-        verify(telegramBot, times(1)).execute(any(SendMessage.class));
+        verify(telegramBot, times(1)).execute(message);
     }
 
     @Test
-    @DisplayName("Проверяет, что при вызове метода handle класса TransportationRecommendationHandler " +
-            "в состоянии \"Рекомендации по транспортировке кошки\" отправляется сообщение с текстом " +
+    @DisplayName("Проверяет, что при вызове метода handle класса RefusalReasonHandler " +
+            "в состоянии \"Причины, почему могут отказать и не дать забрать собаку из приюта\" отправляется сообщение с текстом " +
             "в чат с заданным chatId.")
-    public void testTransportationRecommendationHandlerCat() {
+    public void testDocumentsListHandler() {
         Long chatId = 123L;
 
-        SendMessage message = new SendMessage(chatId, TRANSPORTATION_RECOMMENDATION_CAT_TEXT);
+        SendMessage message = new SendMessage(chatId, REFUSAL_REASON_TEXT);
         telegramBot.execute(message);
         verify(telegramBot, times(1)).execute(any(SendMessage.class));
     }
 
     @Test
     @DisplayName("Проверяет, что при вызове метода handle " +
-            "класса TransportationRecommendationHandler в состоянии \"Стоп\" вызывается метод offerToStart " +
+            "класса RefusalReasonHandler в состоянии \"Стоп\" вызывается метод offerToStart " +
             "класса CommonUtils с заданным chatId.")
     public void testHandleWhenCurrentStateIsStop() {
         when(chatStateService.getCurrentStateByChatId(123L)).thenReturn(BotCommand.STOP);
 
-        transportationRecommendationHandler.handle(update);
+        refusalReasonHandler.handle(update);
 
         verify(commonUtils).offerToStart(123L);
     }
 
     @Test
-    @DisplayName("Проверяет, что при вызове метода handle класса TransportationRecommendationHandler \" +\n"
+    @DisplayName("Проверяет, что при вызове метода handle класса RefusalReasonHandler \" +\n"
             + "в состоянии \\\"Назад\\\" вызывается метод sendInvalidCommandResponse класса CommonUtils с заданным chatId.")
     public void testHandleWhenCurrentStateIsBack() {
         when(chatStateService.getCurrentStateByChatId(123L)).thenReturn(BotCommand.BACK);
 
-        transportationRecommendationHandler.handle(update);
+        refusalReasonHandler.handle(update);
 
         verify(commonUtils).sendInvalidCommandResponse(123L);
     }
 
     @Test
-    @DisplayName("Проверяет, что метод getCommand класса TransportationRecommendationHandler возвращает правильную команду BotCommand.TRANSPORTATION_RECOMMENDATION")
+    @DisplayName("Проверяет, что метод getCommand класса RefusalReasonHandler возвращает правильную команду BotCommand.REFUSAL_REASON.")
     public void testGetCommand() {
-        BotCommand expectedCommand = BotCommand.TRANSPORTATION_RECOMMENDATION;
-        BotCommand actualCommand = transportationRecommendationHandler.getCommand();
+        BotCommand expectedCommand = BotCommand.REFUSAL_REASON;
+        BotCommand actualCommand = refusalReasonHandler.getCommand();
         assertEquals(expectedCommand, actualCommand);
     }
 }
